requirements.txt
Django==3.1.1
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.1